make/env.cmake: Use find_program to find programs from the environment
When USE_CONDA is false, use the environment variables to get the path
to programs and fall back to using find_program to get the full path to
programs. If either fails generate an error message.

Also remove the setting of the value to just the binary name as this
causes find_program to treat that as the value and a path to the binary
that is not a full path will cause issues with all targets that use the
value as a dependency.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>