xml: Improve xslt transform.
Continue to improve the `xmlsort.xsl` script to be a tool which merges a
modified FPGA architecture.

Currently it also converts a more composable format and converts it into
the format that VtR accepts. Eventually the hope is that upstream VtR
will accept this composable format in the future. See;
https://github.com/verilog-to-routing/vtr-verilog-to-routing/issues/272

Signed-off-by: Tim 'mithro' Ansell <me@mith.ro>