Add site pin delay to connection box annotation.
This is important because the generic delay matrix doesn't account for
per site pin delays.  By adding an annotation of the portion of the IPIN
that varies dramatically (and varies at the end of the route and is
unaviodable), it enables the connection box lookahead to properly
account for this delay.

See https://github.com/verilog-to-routing/vtr-verilog-to-routing/issues/1027
for more details and examples.

Signed-off-by: Keith Rothman <537074+litghost@users.noreply.github.com>