Change update_tools to avoid requiring force pushes all the time.
The previous version of update_tools maintained a simplified history by
force pushing overtop of master+wip and master+wip-next.  This results
in github PRs not working correctly, and continuously orphaning trees.

This approach avoids the force push by making a revert commit to
master+wip, and then merging with master.  This achieves the original
goal of keeping the history sane, but avoids orphaning the old
master+wip integration point.

See https://github.com/SymbiFlow/vtr-verilog-to-routing/pull/319 as an
example PR with this new methodology.

Signed-off-by: Keith Rothman <537074+litghost@users.noreply.github.com>