Refactor CMake system to enable ROI-less xc7 graphs that share arches.
Previously some xc7 properties were attached to devices, but belonged to
the board.  This is now fixed.

Moving the PINMAP to the board is less obvious, but I think late binding
the PINMAP allows for less confusion, and more flexiblity.  This does
result in some data duplication, but not a lot.

Signed-off-by: Keith Rothman <537074+litghost@users.noreply.github.com>