Add ability to copy header files to add_fpga_target
This commit extends the SymbiFlow build system,
by introducing an INCLUDES option in the ADD_FPGA_TARGET function.
This option can be used to copy all the necessary files to the build
directory. The files will not be passed to the synthesis tool.
This might be used for the header or vmem files that should be placed
in the build directory together with all the source files.

Signed-off-by: Robert Winkler <rwinkler@antmicro.com>